created the architecture for calendar page
