+++ src/App.css
@@ -1,0 +1,20 @@
+* {
+    margin: 0;
+    padding: 0;
+}
+
+#root {
+    display: flex;
+    height: 100vw;
+}
+
+main {
+    width: 100%;
+    padding: 2%;
+}
+
+#test {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+}
