+++ src/App.jsx
@@ -1,14 +1,13 @@
 import "./App.css";
 import "react-big-calendar/lib/css/react-big-calendar.css";
 import CalendarPage from "./pages/CalendarPage";
+import { Menu } from "./components/styled/Menu.styled"
 
 function App() {
   
   return (
     <>
-      <header>
-        <h1>Leons app</h1>
-      </header>
+      <Menu />
       <main>
         <CalendarPage />
       </main>
