+++ src/components/CurrentDayContainer.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { StyledCurrentDayContainer } from "./styled/CurrentDayContainer.styled";
+
+const CurrentDayContainer = () => {
+  return (
+    <StyledCurrentDayContainer>
+      <h2>Current date and day</h2>
+    </StyledCurrentDayContainer>
+  );
+};
+
+export default CurrentDayContainer;
