+++ src/components/styled/CalendarPage.styled.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+import React from "react";
+
+export const StyledCalendarPage = styled.div`
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+`;
