+++ src/components/styled/CurrentDayContainer.styled.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const StyledCurrentDayContainer = styled.div`
+width: 30%;
+height: 200px;
+`
