--- src/components/styled/CurrentDayContainer.styyled.js
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-export const CurrentDayContainer = styled.div`
-width: 30%;
-`
