+++ src/components/styled/Menu.styled.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const Menu = styled.menu`
+    background-color: black;
+    width: 5%;
+`
